requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
finviz==1.4.0
finvizfinance==0.9.4
pandas==1.2.4
user_agent==0.1.9
quickfs==0.1.1
python-dotenv==0.19.0
